Started moving map over to tilemap, wierd craps happening...
when I flip a tile thoug, it ends up flipping the graphic so that you're looking through the back of it, so the actual texture is on the opposite side of the grid object, not sure what's going on, as it was working earlier. Use the grass tileset, not the other one, I did some work on colliders so they macth the tiles better.